Recover Secret modules and files refactoring
