+++ client/recover_secret/data_structures.rs
