+++ client/recover_secret/mod.rs
@@ -1,0 +1,3 @@
+mod utils;
+pub mod data_structures;
+pub mod recover_secret_challenge;
